frontend/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
streamlit
requests
python-dotenv